Adding initial classes
Adding initial classes for Waypoints, Routes, collections of Waypoints and Routes and a union of Waypoints and Routes (Network).

To test these classes I’ve added some test data (12 waypoints and 3 routes) using a temporary basic schema with a subclassed Network class to load this test data and create Waypoint/Route instances and collections. To verify exports from this project in the future, a QGIS project is setup using a GeoJSON version of the same test data.

In time that test class and data will be replaced by something more real/realistic.