Adding FPL validation
This adds the Garmin FPL XSD schema as a resource file (which we’ve had to modify to address issues with regular expressions, and to allow us to export FPLs without waypoint tables in route files).

Validation is performed via the xmllint command line tool via sub-process.

See #4 for details about issues implementing this validation.