Aligning route waypoint model with waypoint
As waypoint now contains a description property, it’s confusing to have a description (which is a comment) in route waypoints.

It’s also confusing to have both a general comment for a waypoint, and a comment that’s route specific but I think that’s less confusing that description/comment.