Fixing comment regression in GPX files
Route waypoints in GPX files should include the comment from each corresponding Waypoint to ensure reference files don’t change.

Note there is a lack of integrity and consistency by doing this. Integirty because only the comment is included, not the other fields that make up a Waypoint description. Consistency because these other descriptive elements are included in Waypoints within the GPX, and no comment is included in Route CSV files (which aren’t used and so not tested properly).

This should be addressed in the future.